+++ plugins/postcss-double-position-gradients/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Double Position Gradients
 
+### Unreleased (patch)
+
+- Add typescript support
+- Fix `at` keyword with `at 20px 20px` being interpreted as a double position color stop.
+
 ### 3.1.0 (February 15, 2022)
 
 - Ignore values in relevant `@supports` rules.